Curate public API surface ahead of 1.0.0
- Require surrealdb peer dependency >=2.0.0 (drop alpha pre-release)
- Update README requirements and stability framing toward 1.0.0
- Hide internal plumbing from the package entry point: intoWorkable,
  isWorkable, sanitizeWorkable, workableGet, createVariableStore,
  traversableRow, and the standalone function builders (standaloneFn,
  standaloneConst, extractContext) now moved to functions/standalone/internal
- Re-export ./utils explicitly so only the documented escape hatch
  (displayContext/__display) and signature-bearing types stay public
- Document why the set_/type_ families carry trailing underscores

Public surface: 100 symbols (was 109). Type-check, unit tests, and lint pass.
