+++ src/functions/standalone/count.ts
@@ -1,6 +1,6 @@
 import { t } from "../../types";
 import type { Workable, WorkableContext } from "../../utils";
-import { type ContextSource, standaloneFn } from "./index";
+import { type ContextSource, standaloneFn } from "./internal";
 
 /** count() - counts a row, or counts truthy values if a value is given */
 export function count<C extends WorkableContext>(
