+++ src/functions/standalone/index.ts
@@ -1,72 +1,9 @@
-import type { AbstractType } from "../../types";
-import {
-	__ctx,
-	__display,
-	__type,
-	isWorkable,
-	type Workable,
-	type WorkableContext,
-} from "../../utils";
-import { type Actionable, actionable } from "../../utils/actionable";
-
-/**
- * Context source for standalone functions.
- * Accepts either a WorkableContext directly or a Workable to extract context from.
- */
-export type ContextSource<C extends WorkableContext = WorkableContext> =
-	| C
-	| Workable<C>;
-
-/**
- * Extract a WorkableContext from either a raw context or a Workable.
- */
-export function extractContext<C extends WorkableContext>(
-	source: ContextSource<C>,
-): C {
-	if (isWorkable(source)) {
-		return source[__ctx];
-	}
-	return source as C;
-}
-
-/**
- * Create a standalone database function that generates SurrealQL like `fn_name(param1, param2)`.
- * Unlike databaseFunction in ../utils.ts, this doesn't require a pre-existing Workable context binding.
- */
-export function standaloneFn<C extends WorkableContext, T extends AbstractType>(
-	source: ContextSource<C>,
-	type: T,
-	fn: string,
-	...params: Workable<C>[]
-): Actionable<C, T> {
-	const ctx = extractContext(source);
-	return actionable({
-		[__ctx]: ctx,
-		[__type]: type,
-		[__display](ctx) {
-			const vars = params.map((p) => p[__display](ctx)).join(", ");
-			return vars ? `${fn}(${vars})` : `${fn}()`;
-		},
-	});
-}
-
-/**
- * Create a standalone constant reference (no parentheses).
- * Generates SurrealQL like `math::pi` or `math::e`.
- */
-export function standaloneConst<
-	C extends WorkableContext,
-	T extends AbstractType,
->(source: ContextSource<C>, type: T, name: string): Actionable<C, T> {
-	const ctx = extractContext(source);
-	return actionable({
-		[__ctx]: ctx,
-		[__type]: type,
-		[__display]() {
-			return name;
-		},
-	});
-}
+// Standalone function families.
+//
+// The internal builders (`standaloneFn`, `standaloneConst`, `extractContext`)
+// live in ./internal and are intentionally NOT re-exported — they are
+// implementation plumbing. `ContextSource` is re-exported because it appears in
+// the public signatures of the families below.
 
 export * from "./and";
 export * from "./bytes";
@@ -76,6 +13,7 @@ export * from "./duration";
 export * from "./encoding";
 export * from "./geo";
 export * from "./http";
+export type { ContextSource } from "./internal";
 export * from "./math";
 export * from "./meta";
 export * from "./not";
