+++ src/functions/standalone/meta.ts
@@ -1,6 +1,6 @@
 import { t } from "../../types";
 import type { Workable, WorkableContext } from "../../utils";
-import { standaloneFn } from "./index";
+import { standaloneFn } from "./internal";
 
 export const meta = {
 	id<C extends WorkableContext>(value: Workable<C>) {
