+++ src/functions/standalone/object.ts
@@ -1,6 +1,6 @@
 import { t } from "../../types";
 import type { Workable, WorkableContext } from "../../utils";
-import { standaloneFn } from "./index";
+import { standaloneFn } from "./internal";
 
 export const object = {
 	entries<C extends WorkableContext>(value: Workable<C>) {
