+++ src/functions/standalone/session.ts
@@ -1,6 +1,6 @@
 import { t } from "../../types";
 import type { WorkableContext } from "../../utils";
-import { type ContextSource, standaloneFn } from "./index";
+import { type ContextSource, standaloneFn } from "./internal";
 
 export const session = {
 	ac<C extends WorkableContext>(source: ContextSource<C>) {
